Increased delay in test functions
The functions were checking if previous requests are being cancelled when new ones are sent. In case the CI server has a case of the Mondays and is slow, it lead to a failure. The delay was increaed from 10 to 100ms, which should fix that. But it's more of a hack than a real solution.